requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
watchfiles>=0.21.0
anyio==4.3.0
flask==3.0.0
requests==2.31.0
urllib3<3
charset_normalizer<4
sqlalchemy==2.0.38
pandas==2.2.3
numpy==2.2.3
plotly==6.0.0
matplotlib==3.10.1
langgraph==0.3.21
langchain-core==0.3.44
langchain-groq
python-multipart==0.0.20
reportlab==4.2.5
slowapi==0.1.9
tenacity==8.2.3
bleach==6.1.0
